Switch invite link format for OpSec
Now users get a unique link for each invitation, rather than a URL
composed of predictable components.

Previously: /view/EVENT_ID/USER_ID
New: /view/RSVP_ID

With this update, if a friend sends you their unique invite, you can
edit it. However, you will not be able to copy their ID and use it to
RSVP to other events on the same server. This is an acceptable threat
profile for our purposes.